move ttl_to_seconds from cache_utils to runtime_utils
