+++ lib/streamlit/runtime/caching/cache_resource_api.py
@@ -35,7 +35,6 @@ from streamlit.runtime.caching.cache_utils import (
     Cache,
     CachedFuncInfo,
     make_cached_func_wrapper,
-    ttl_to_seconds,
 )
 from streamlit.runtime.caching.cached_message_replay import (
     CachedMessageReplayContext,
@@ -46,6 +45,7 @@ from streamlit.runtime.caching.cached_message_replay import (
 )
 from streamlit.runtime.caching.hashing import HashFuncsDict
 from streamlit.runtime.metrics_util import gather_metrics
+from streamlit.runtime.runtime_util import ttl_to_seconds
 from streamlit.runtime.scriptrunner.script_run_context import get_script_run_ctx
 from streamlit.runtime.stats import CacheStat, CacheStatsProvider, group_stats
