+++ lib/streamlit/runtime/runtime_util.py
@@ -16,11 +16,14 @@
 
 from __future__ import annotations
 
-from typing import Any
+import math
+from datetime import timedelta
+from typing import Any, Literal, overload
 
 from streamlit import config
 from streamlit.errors import MarkdownFormattedException
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
+from streamlit.runtime.caching.cache_errors import BadTTLStringError
 from streamlit.runtime.forward_msg_cache import populate_hash_if_needed
 
 
@@ -60,6 +63,45 @@ def is_cacheable_msg(msg: ForwardMsg) -> bool:
         # Some message types never get cached
         return False
     return msg.ByteSize() >= int(config.get_option("global.minCachedMessageSize"))
+
+
+@overload
+def ttl_to_seconds(
+    ttl: float | timedelta | str | None, *, coerce_none_to_inf: Literal[False]
+) -> float | None:
+    ...
+
+
+@overload
+def ttl_to_seconds(ttl: float | timedelta | str | None) -> float:
+    ...
+
+
+def ttl_to_seconds(
+    ttl: float | timedelta | str | None, *, coerce_none_to_inf: bool = True
+) -> float | None:
+    """
+    Convert a ttl value to a float representing "number of seconds".
+    """
+    if coerce_none_to_inf and ttl is None:
+        return math.inf
+    if isinstance(ttl, timedelta):
+        return ttl.total_seconds()
+    if isinstance(ttl, str):
+        import numpy as np
+        import pandas as pd
+
+        try:
+            out: float = pd.Timedelta(ttl).total_seconds()
+        except ValueError as ex:
+            raise BadTTLStringError(ttl) from ex
+
+        if np.isnan(out):
+            raise BadTTLStringError(ttl)
+
+        return out
+
+    return ttl
 
 
 def serialize_forward_msg(msg: ForwardMsg) -> bytes:
